% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportForecolte.R
\name{ImportForecolte}
\alias{ImportForecolte}
\title{Import data.}
\usage{
ImportForecolte(rep = getwd())
}
\arguments{
\item{rep}{répertoire conteant les données}
}
\description{
Import des informations nécessaires au calcul des volumes mobilisables.
}
\examples{
rep = "/Users/maxbruciamacchie/pCloudSync/EnCours/Possibilités/Data/Foret1"
ImportForestCrop(rep)

}
\author{
Bruciamacchie Max
}
