% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrigeageMnhFonc.R
\name{KrigeageMnhFonc}
\alias{KrigeageMnhFonc}
\title{Fonction Krigeage avec MNH}
\usage{
KrigeageMnhFonc(shpPerim, shpPlac, mnh, pas = 25, idvar = "Gha")
}
\description{
Fonction krigeage pour conversion des données ponctuelles. Prise en compte d'un MNH
}
\examples{
k <- KrigeageMnhFonc(perim, plac, mnh15, idvar="Vha")

}
\author{
Bruciamacchie Max
}
